fix: extract flat-providers into core folder to exclude circular dependency
